+++ pages/vendorsearch/[vendorsearch].js
@@ -4,7 +4,7 @@ import Head from 'next/head';
 import ProductCards from '../../components/Product/ProductCards';
 import CategorySelect from '../../components/Categories/CategorySelect';
 import { vendors } from '../../VendorList/VendorList';
-import { checkProductDiscounts } from '../../components/utils';
+import { checkProductDiscounts } from '../../utils/sales';
 const JSONBig = require('json-bigint');
 const { Client, Environment } = require('square');
 import { useSelector } from 'react-redux';
